Fix perspective
When running this in chrome the rectangles didn't move up and
down the y-axis, only circling in the x axis. I don't fully
understand what this code does, but it seems applying z_delta
to the z-axis doesn't yield the (I presume) intended effect
of the wings moving up and down. Changing this to instead
apply to the y-axis yields a good result.

The factor is something I chose because It Looked Good on My
Machine (tm) but it can probably be something smarter.