Fix perspective transform.
Marcus Stade noticed that the z-component of the transform was having no effect; this was because the perpspective transform on the camera was getting truncated. This change sets "-webkit-transform-style: preserve-3d" on the camera so that the transform doesn't get flattened. It also shrinks the magnitude of the z translation a bit and adds a rotation to the camera.