Up the z amplitude a bit and lower the camera position
It looked a bit squashed with the lower amp so I put it back
up. But I did so while making it view dependent, so that it
doesn't shoot up in the sky if the window is resized. It also
didn't look positioned quite right so I dropped
the camera a bit by putting it at `top: 25%`.

I also changed the rectangle width/height to use `vmin`
instead of percentages. This ensures that they stay square,
regardless of the width/height of the window.

Finally, I changed the camera perspective value to give a more
orthographic look, which made it more alike the recollection
I have of the original.

Man I love this thing.