Initial work on firefox compatibility
I basically yanked out the WebkitCSSTransform usage and
replaced with simple array manipulation. It seems to work as
it used to, I don't *think* I screwed anything up.

The visual fidelity of this isn't quite on par with the Chrome
version, so there's some tweaking to be done. Also, the perf
isn't great so I'll look into optimizing it some.

This is work in progress.