Changed to reuse DOM elements instead of creating new ones
This update changes the wings to update in place, rather than
creating new instances. Crucially, this means the DOM elements
are reused rather than replaced every frame update.

This seems to have a significant performance boost in Firefox;
performance in Chrome is the same or better, but not nearly as
obvious. I have no measurements to back these claims up,
just eyeballing. Still, it's a very noticable speedup I think.

There are still hickups, so further work is necessary.